# openapi.preview.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Telegram Cleaner Bot API",
    "summary": "Telegram 群组清理与审核管理 API",
    "description": "用于 Telegram 群组治理的后端 API。\n\n- 提供群组、规则、处置记录三类核心资源管理能力\n- 统一返回结构化错误体，便于前端与 AI 调用端稳定解析\n- 列表接口均采用 limit/offset 分页语义\n\n鉴权约定（前端调用必读）\n- 受保护路由：/groups、/rules、/moderation（必须带 X-API-Key）\n- 公开路由：/、/health（无需鉴权）\n- 请求头示例：X-API-Key: your_api_secret_key\n- curl 示例：curl -H \"X-API-Key: your_api_secret_key\" http://localhost:8000/groups?limit=10&offset=0\n- 鉴权失败：返回 401，错误码 UNAUTHORIZED。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/groups": {
      "post": {
        "tags": [
          "groups"
        ],
        "summary": "创建群组",
        "description": "创建一个新的群组配置记录，返回创建后的完整群组信息。",
        "operationId": "create_group",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupRead"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "鉴权失败：缺失或无效的 X-API-Key 请求头",
            "content": {
              "application/json": {
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "API 密钥无效",
                  "request_id": "a29f13f9a8d24e6da3a6c4d9ce80f641",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/groups/1"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "groups"
        ],
        "summary": "查询群组列表",
        "description": "按分页条件与启用状态筛选群组列表。",
        "operationId": "list_groups",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupListResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "鉴权失败：缺失或无效的 X-API-Key 请求头",
            "content": {
              "application/json": {
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "API 密钥无效",
                  "request_id": "a29f13f9a8d24e6da3a6c4d9ce80f641",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/groups/1"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/groups/{group_id}": {
      "get": {
        "tags": [
          "groups"
        ],
        "summary": "查询群组详情",
        "description": "按群组主键 ID 查询单个群组详情。",
        "operationId": "get_group_by_id",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "群组主键 ID",
              "title": "Group Id"
            },
            "description": "群组主键 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupRead"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "鉴权失败：缺失或无效的 X-API-Key 请求头",
            "content": {
              "application/json": {
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "API 密钥无效",
                  "request_id": "a29f13f9a8d24e6da3a6c4d9ce80f641",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/groups/1"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "groups"
        ],
        "summary": "更新群组",
        "description": "按群组主键 ID 局部更新群组配置。",
        "operationId": "update_group_by_id",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "群组主键 ID",
              "title": "Group Id"
            },
            "description": "群组主键 ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupRead"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "鉴权失败：缺失或无效的 X-API-Key 请求头",
            "content": {
              "application/json": {
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "API 密钥无效",
                  "request_id": "a29f13f9a8d24e6da3a6c4d9ce80f641",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/groups/1"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "groups"
        ],
        "summary": "删除群组",
        "description": "按群组主键 ID 删除群组。删除成功后返回 204，无响应体。",
        "operationId": "delete_group_by_id",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "群组主键 ID",
              "title": "Group Id"
            },
            "description": "群组主键 ID"
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功，无响应体"
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "鉴权失败：缺失或无效的 X-API-Key 请求头",
            "content": {
              "application/json": {
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "API 密钥无效",
                  "request_id": "a29f13f9a8d24e6da3a6c4d9ce80f641",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/groups/1"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/groups/{group_id}/authorization": {
      "patch": {
        "tags": [
          "groups"
        ],
        "summary": "更新群组授权状态",
        "description": "按群组主键 ID 更新群组授权状态。",
        "operationId": "update_group_authorization_by_id",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "群组主键 ID",
              "title": "Group Id"
            },
            "description": "群组主键 ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupAuthorizationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupRead"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "鉴权失败：缺失或无效的 X-API-Key 请求头",
            "content": {
              "application/json": {
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "API 密钥无效",
                  "request_id": "a29f13f9a8d24e6da3a6c4d9ce80f641",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/groups/1"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/groups/{group_id}/settings": {
      "put": {
        "tags": [
          "groups"
        ],
        "summary": "全量覆盖群组设置",
        "description": "按群组主键 ID 全量覆盖群组设置。",
        "operationId": "replace_group_settings_by_id",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "群组主键 ID",
              "title": "Group Id"
            },
            "description": "群组主键 ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupSettingsReplace"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupRead"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "鉴权失败：缺失或无效的 X-API-Key 请求头",
            "content": {
              "application/json": {
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "API 密钥无效",
                  "request_id": "a29f13f9a8d24e6da3a6c4d9ce80f641",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/groups/1"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在",
                  "request_id": "c6f7f8c4a0c346f28fe5d4c0fcab0d72",
                  "timestamp": "2026-05-04T08:00:00Z",
                  "path": "/api/groups/999"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GroupAuthorizationUpdate": {
        "properties": {
          "is_authorized": {
            "type": "boolean",
            "title": "Is Authorized",
            "description": "是否已授权使用机器人"
          }
        },
        "type": "object",
        "required": [
          "is_authorized"
        ],
        "title": "GroupAuthorizationUpdate",
        "description": "更新群组授权状态请求。"
      },
      "GroupChatType": {
        "type": "string",
        "enum": [
          "group",
          "supergroup"
        ],
        "title": "GroupChatType",
        "description": "Telegram 群会话类型。"
      },
      "GroupCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "title": "Title",
            "description": "群组名称"
          },
          "username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Username",
            "description": "群组公开用户名"
          },
          "chat_type": {
            "$ref": "#/components/schemas/GroupChatType",
            "description": "群组类型",
            "default": "supergroup"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "群组描述"
          },
          "owner_telegram_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Telegram User Id",
            "description": "群主 Telegram 用户 ID"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "是否启用清理",
            "default": true
          },
          "settings": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroupSettings"
              },
              {
                "type": "null"
              }
            ],
            "description": "群组维度配置"
          },
          "bot_permissions": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BotPermissions"
              },
              {
                "type": "null"
              }
            ],
            "description": "机器人在群里的权限配置"
          },
          "telegram_group_id": {
            "type": "integer",
            "title": "Telegram Group Id",
            "description": "Telegram 群组 ID"
          }
        },
        "type": "object",
        "required": [
          "title",
          "telegram_group_id"
        ],
        "title": "GroupCreate",
        "description": "创建群组请求。"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "description": "服务 API 鉴权密钥。\n\n受保护路由：/groups、/rules、/moderation。\n公开路由：/、/health。\n\n请求头格式：X-API-Key: <API_SECRET_KEY>\n\n若 API_SECRET_KEY 为空，服务启动时会随机生成一个进程内有效密钥（重启后变化）。",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}